Update CatList.php
Had a case where in my 3.9 install, some posts have a YouTube identifier as the post thumbnail ID, so I added a case to check for and handle that. I'm not sure if that's normal or whether my YouTube importer plugin is doing it wrong, but the default twenty fourteen theme handles the thumbnails fine.