New Attribute 'no_post_titles' for Shortocode 'catlist'
The new attribute 'no_post_titles' is added to shortcode 'catlist',
which can have two values: 'yes' or 'no'. 'no' is the default value,
which will lead to the normal shortcode behavior as in the original
version of the plugin. 'yes' will omit the titles of the posts in
the post list.

This new option is intended for use with 'content=yes', in which
case it will basically include the post text directly without having
a separate title per entry. This allows for building bibliographies
with one literature entry per post and displaying them as lists, as
done for example on the test page
  http://thomasweise.apps-1and1.net/publications/all
(which will eventually be moved elsewhere).

The modification of the plugin is really minimal. Besides adding
the new attribute with default value 'no' in file list_cat_posts.php,
there is only one if-then-else in lcp-catlistdisplayer.php.

The documentation file readme.txt has been extended with a sentence
describing this attribute.